Add tab for map view and list view for Ns local units
