+++ app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/LocalUnitsMap/i18n.json
@@ -1,0 +1,10 @@
+{
+    "namespace": "nationalSocietyLocalUnitsMap",
+    "strings": {
+        "localUnitDetailAddress": "Address",
+        "localUnitDetailPhoneNumber": "Phone Number",
+        "localUnitDetailLastUpdate": "Last Updated",
+        "localUnitDetailFocalPerson": "Focal Person",
+        "localUnitDetailEmail": "Email"
+    }
+}
