+++ app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/LocalUnitsMap/index.tsx
@@ -7,23 +7,14 @@ import { useOutletContext } from 'react-router-dom';
 import {
     LocationIcon,
     MailIcon,
-    SearchLineIcon,
 } from '@ifrc-go/icons';
 import {
-    Button,
     Container,
     List,
-    SelectInput,
-    TextInput,
     TextOutput,
 } from '@ifrc-go/ui';
 import { useTranslation } from '@ifrc-go/ui/hooks';
 import {
-    stringLabelSelector,
-    stringNameSelector,
-} from '@ifrc-go/ui/utils';
-import {
-    _cs,
     isDefined,
     isNotDefined,
     isTruthyString,
@@ -44,22 +35,18 @@ import BaseMap from '#components/domain/BaseMap';
 import Link, { type Props as LinkProps } from '#components/Link';
 import MapContainerWithDisclaimer from '#components/MapContainerWithDisclaimer';
 import MapPopup from '#components/MapPopup';
-import { adminUrl } from '#config';
-import { type components } from '#generated/types';
-import useAuth from '#hooks/domain/useAuth';
-import useFilterState from '#hooks/useFilterState';
 import {
     COLOR_RED,
     DEFAULT_MAP_PADDING,
     DURATION_MAP_ZOOM,
 } from '#utils/constants';
 import { type CountryOutletContext } from '#utils/outletContext';
-import { resolveUrl } from '#utils/resolveUrl';
-import { useRequest } from '#utils/restRequest';
+import { type GoApiResponse } from '#utils/restRequest';
 
 import i18n from './i18n.json';
 import styles from './styles.module.css';
 
+type GetLocalUnitResponseType = GoApiResponse<'/api/v2/local-units/'>;
 const basePointPaint: CirclePaint = {
     'circle-radius': 5,
     'circle-color': COLOR_RED,
@@ -73,38 +60,29 @@ const basePointLayerOptions: Omit<CircleLayer, 'id'> = {
     type: 'circle',
     paint: basePointPaint,
 };
+
 const sourceOption: mapboxgl.GeoJSONSourceRaw = {
     type: 'geojson',
 };
-
-type LocalUnitType = components<'read'>['schemas']['LocalUnitType'];
-
-const localUnitCodeSelector = (localUnit: LocalUnitType) => localUnit.code;
-
-interface Validation {
-    label: string;
-}
 
 interface ClickedPoint {
     id: string;
     lngLat: mapboxgl.LngLatLike;
 }
 
-interface Props {
-    className?: string;
-}
-
 function emailKeySelector(email: string) {
     return email;
 }
 
-function NationalSocietyLocalUnitsMap(props: Props) {
-    const {
-        className,
-    } = props;
+interface Props {
+    localUnitListResponse?: GetLocalUnitResponseType;
+}
+
+function LocalUnitsMap(props: Props) {
+    const { localUnitListResponse } = props;
+    const { countryResponse } = useOutletContext<CountryOutletContext>();
 
     const strings = useTranslation(i18n);
-    const { countryId, countryResponse } = useOutletContext<CountryOutletContext>();
     const [
         clickedPointProperties,
         setClickedPointProperties,
@@ -113,45 +91,6 @@ function NationalSocietyLocalUnitsMap(props: Props) {
     const countryBounds = useMemo(() => (
         countryResponse ? getBbox(countryResponse.bbox) : undefined
     ), [countryResponse]);
-
-    const { isAuthenticated } = useAuth();
-
-    const {
-        rawFilter,
-        filter,
-        filtered,
-        setFilter,
-        setFilterField,
-    } = useFilterState<{
-        type?: number;
-        search?: string;
-        isValidated?: string;
-    }>({
-        filter: {},
-        pageSize: 9999,
-    });
-
-    const {
-        response: localUnitListResponse,
-    } = useRequest({
-        skip: isNotDefined(countryResponse?.iso3),
-        url: '/api/v2/local-units/',
-        query: {
-            limit: 9999,
-            type__code: filter.type,
-            validated: isDefined(filter.isValidated)
-                ? filter.isValidated === strings.validated : undefined,
-            search: filter.search,
-            country__iso3: isDefined(countryResponse?.iso3) ? countryResponse?.iso3 : undefined,
-        },
-    });
-
-    const {
-        response: localUnitsOptionsResponse,
-        pending: localUnitsOptionsResponsePending,
-    } = useRequest({
-        url: '/api/v2/local-units/options/',
-    });
 
     const localUnitsGeoJson = useMemo((): GeoJSON.FeatureCollection<GeoJSON.Geometry> => ({
         type: 'FeatureCollection' as const,
@@ -203,15 +142,6 @@ function NationalSocietyLocalUnitsMap(props: Props) {
         [clickedPointProperties, localUnitListResponse?.results],
     );
 
-    const validationOptions: Validation[] = useMemo(() => ([
-        {
-            label: strings.validated,
-        },
-        {
-            label: strings.notValidated,
-        },
-    ]), [strings.validated, strings.notValidated]);
-
     const handlePointClick = useCallback(
         (feature: mapboxgl.MapboxGeoJSONFeature, lngLat: mapboxgl.LngLat) => {
             setClickedPointProperties({
@@ -230,13 +160,6 @@ function NationalSocietyLocalUnitsMap(props: Props) {
         [setClickedPointProperties],
     );
 
-    const handleClearFilter = useCallback(
-        () => {
-            setFilter({});
-        },
-        [setFilter],
-    );
-
     const emailRendererParams = useCallback(
         (_: string, email: string): LinkProps => ({
             className: styles.email,
@@ -249,65 +172,7 @@ function NationalSocietyLocalUnitsMap(props: Props) {
     );
 
     return (
-        <Container
-            className={_cs(styles.nationalSocietyLocalUnitsMap, className)}
-            heading={strings.localUnitsMapTitle}
-            childrenContainerClassName={styles.content}
-            withGridViewInFilter
-            withHeaderBorder
-            actions={isAuthenticated && (
-                <Link
-                    external
-                    href={resolveUrl(adminUrl, `local_units/localunit/?country=${countryId}`)}
-                    variant="secondary"
-                >
-                    {strings.editLocalUnitLink}
-                </Link>
-            )}
-            filters={(
-                <>
-                    <SelectInput
-                        placeholder={strings.localUnitsFilterTypePlaceholder}
-                        label={strings.localUnitsFilterTypeLabel}
-                        name="type"
-                        value={rawFilter.type}
-                        onChange={setFilterField}
-                        keySelector={localUnitCodeSelector}
-                        labelSelector={stringNameSelector}
-                        disabled={localUnitsOptionsResponsePending}
-                        options={localUnitsOptionsResponse?.type}
-                    />
-                    <SelectInput
-                        placeholder={strings.localUnitsFilterValidatedPlaceholder}
-                        label={strings.localUnitsFilterValidatedLabel}
-                        name="isValidated"
-                        value={rawFilter.isValidated}
-                        onChange={setFilterField}
-                        keySelector={stringLabelSelector}
-                        labelSelector={stringLabelSelector}
-                        options={validationOptions}
-                    />
-                    <TextInput
-                        name="search"
-                        label={strings.localUnitsFilterSearchLabel}
-                        placeholder={strings.localUnitsFilterSearchPlaceholderLabel}
-                        value={rawFilter.search}
-                        onChange={setFilterField}
-                        icons={<SearchLineIcon />}
-                    />
-                    <div className={styles.clearButton}>
-                        <Button
-                            name={undefined}
-                            variant="secondary"
-                            onClick={handleClearFilter}
-                            disabled={!filtered}
-                        >
-                            {strings.localUnitsFilterClear}
-                        </Button>
-                    </div>
-                </>
-            )}
-        >
+        <>
             <BaseMap
                 baseLayers={(
                     <MapLayer
@@ -398,54 +263,60 @@ function NationalSocietyLocalUnitsMap(props: Props) {
                 )}
             </BaseMap>
             {isDefined(countryResponse)
-                && (isDefined(countryResponse.address_1) || isDefined(countryResponse.emails)) && (
-                <Container
-                    className={styles.mapDetail}
-                    childrenContainerClassName={styles.infoContainer}
-                >
-                    {isDefined(countryResponse) && isDefined(countryResponse.address_1) && (
-                        <TextOutput
-                            className={styles.info}
-                            labelClassName={styles.label}
-                            label={(
-                                <LocationIcon className={styles.icon} />
-                            )}
-                            withoutLabelColon
-                            value={(
-                                <>
-                                    <div>{countryResponse.address_1}</div>
-                                    <div>{countryResponse.address_2}</div>
-                                </>
-                            )}
-                        />
+                && (isDefined(countryResponse.address_1)
+                    || isDefined(countryResponse.emails))
+                    && (
+                        <Container
+                            className={styles.mapDetail}
+                            childrenContainerClassName={styles.infoContainer}
+                        >
+                            {isDefined(countryResponse)
+                                && isDefined(countryResponse.address_1)
+                                && (
+                                    <TextOutput
+                                        className={styles.info}
+                                        labelClassName={styles.label}
+                                        label={(
+                                            <LocationIcon className={styles.icon} />
+                                        )}
+                                        withoutLabelColon
+                                        value={(
+                                            <>
+                                                <div>{countryResponse.address_1}</div>
+                                                <div>{countryResponse.address_2}</div>
+                                            </>
+                                        )}
+                                    />
+                                )}
+                            {isDefined(countryResponse)
+                                && isDefined(countryResponse.emails)
+                                && (
+                                    <TextOutput
+                                        className={styles.info}
+                                        labelClassName={styles.label}
+                                        label={(
+                                            <MailIcon className={styles.icon} />
+                                        )}
+                                        withoutLabelColon
+                                        value={(
+                                            <List
+                                                data={countryResponse.emails.filter(isDefined)}
+                                                renderer={Link}
+                                                rendererParams={emailRendererParams}
+                                                keySelector={emailKeySelector}
+                                                withoutMessage
+                                                compact
+                                                pending={false}
+                                                errored={false}
+                                                filtered={false}
+                                            />
+                                        )}
+                                    />
+                                )}
+                        </Container>
                     )}
-                    {isDefined(countryResponse) && isDefined(countryResponse.emails) && (
-                        <TextOutput
-                            className={styles.info}
-                            labelClassName={styles.label}
-                            label={(
-                                <MailIcon className={styles.icon} />
-                            )}
-                            withoutLabelColon
-                            value={(
-                                <List
-                                    data={countryResponse.emails.filter(isDefined)}
-                                    renderer={Link}
-                                    rendererParams={emailRendererParams}
-                                    keySelector={emailKeySelector}
-                                    withoutMessage
-                                    compact
-                                    pending={false}
-                                    errored={false}
-                                    filtered={false}
-                                />
-                            )}
-                        />
-                    )}
-                </Container>
-            )}
-        </Container>
+        </>
     );
 }
 
-export default NationalSocietyLocalUnitsMap;
+export default LocalUnitsMap;
