+++ app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/LocalUnitsTable/index.tsx
@@ -1,0 +1,7 @@
+function LocalUnitsTable() {
+    return (
+        <div> Table </div>
+    );
+}
+
+export default LocalUnitsTable;
